adicionado banco de dados para registrar emails enviados
