+++ infopublic_mail/templates/usuario.html
@@ -30,8 +30,7 @@
         {% endif %}
         </div>
         <div class="col-md-6">
-        <table class="table"> 
-            
+        <table class="table">             
             <tbody class="info">
                 {% for entidade in entidades %}
                     <tr class="info">
